minicloud/app/requirements.txt
Flask==3.0.3
PyJWT[crypto]==2.9.0
requests==2.32.3
mysql-connector-python==9.1.0
